Refactor code to use new places service
