+++ apps/web/src/services/store.ts
@@ -1,49 +1,12 @@
-import {
-  configureStore,
-  ThunkAction,
-  Action,
-  createSlice,
-} from "@reduxjs/toolkit";
+import { configureStore, ThunkAction, Action } from "@reduxjs/toolkit";
 import auth from "./auth";
 import { useAppDispatch, useAppSelector } from "./hooks";
 import toastSlice from "./toast/toast.slice";
-
-const initialState: {
-  isOpen: boolean;
-  listId?: string;
-  onClose: () => void;
-  place: google.maps.places.PlaceResult;
-} = {
-  isOpen: false,
-  listId: null,
-  place: null,
-  onClose: null,
-};
-
-const placeModalSlice = createSlice({
-  name: "placeModal",
-  initialState,
-  reducers: {
-    openPlaceModal(state, action) {
-      state.isOpen = true;
-      state.place = action.payload.place;
-      state.listId = action.payload.listId;
-      state.onClose = action.payload.onClose;
-    },
-    closePlaceModal(state) {
-      state.isOpen = false;
-      state.place = null;
-      state.onClose?.();
-      state.onClose = null;
-    },
-  },
-});
-
-export const { closePlaceModal } = placeModalSlice.actions;
+import { placesSlice } from "./places";
 
 export const rootReducer = {
   auth,
-  placeModal: placeModalSlice.reducer,
+  placeModal: placesSlice.reducer,
   toast: toastSlice.reducer,
 };
 
@@ -54,28 +17,6 @@ export const store = configureStore({
 export const getIsLoadingAuth = (state: RootState) => state.auth.isLoadingAuth;
 export const getLoginEmail = (state: RootState) => state.auth.loginEmail;
 export const getIsAuthenticated = (state: RootState) => !!state.auth.user;
-
-export const usePlaceModal = () => {
-  const dispatch = useAppDispatch();
-
-  const openPlaceModal = ({
-    listId,
-    place,
-    onClose,
-  }: {
-    listId?: string;
-    place: google.maps.places.PlaceResult;
-    onClose?: () => void;
-  }) => {
-    return dispatch(
-      placeModalSlice.actions.openPlaceModal({ listId, place, onClose }),
-    );
-  };
-
-  return {
-    openPlaceModal,
-  };
-};
 
 export const useAuth = () => {
   const user = useAppSelector((state: RootState) => state.auth.user);
